feat: implement method in cjsToEsm.cjs
