+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,29 @@
+import * as path from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+import jsonA from './files/a.json' assert {type: "json"};
+import jsonB from './files/b.json' assert {type: "json"};
+import { fileURLToPath } from 'url';
+
+export let unknownObject;
+const random = Math.random();
+if (random > 0.5) {
+    unknownObject = jsonA;
+} else {
+    unknownObject = jsonB;
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+const __filename = fileURLToPath(import.meta.url);
+console.log(`Path to current file is ${__filename}`);
+const __dirname = path.dirname(__filename);
+console.log(`Path to current directory is ${__dirname}`);
+
+export const createMyServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
